application.properties - убрал сваггер проперти new view CreateTestComponent (creteTest) MainPage - editTheme variants OpenApiConfig - openApi to swagger 2 pom.xml - del openapi, add springfox, add markdown ProjectsView.java createProjectButton and more refactoring

+++ src/main/java/org/tms/tms/web/view/ProjectsView.java
@@ -1,10 +1,10 @@
 package org.tms.tms.web.view;
 
-import com.vaadin.flow.component.AttachEvent;
 import com.vaadin.flow.component.ClickEvent;
 import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.dialog.Dialog;
 import com.vaadin.flow.component.formlayout.FormLayout;
 import com.vaadin.flow.component.grid.ItemClickEvent;
@@ -21,13 +21,10 @@ import org.tms.tms.api.ProjectController;
 import org.tms.tms.api.SuiteController;
 import org.tms.tms.api.TestController;
 import org.tms.tms.dao.Project;
-import org.tms.tms.dao.Suite;
-import org.tms.tms.dao.Test;
 import org.tms.tms.dto.ProjectDto;
 import org.vaadin.crudui.crud.impl.GridCrud;
 
-import java.util.Collection;
-
+@CssImport("./styles/shared-styles.css")
 @Route(value = "Projects", layout = MainPage.class)
 @PageTitle("Projects")
 public class ProjectsView extends VerticalLayout {
@@ -42,55 +39,13 @@ public class ProjectsView extends VerticalLayout {
         this.projectController = projectController;
         this.suiteController = suiteController;
         this.testController = testController;
-        Button createProject = new Button();
-        createProject.setIcon(VaadinIcon.PLUS_CIRCLE.create());
-        createProject.setText("Создать проект");
-        add(createProject);
+        add(createProjectButton());
         grid();
-        createProject.addClickListener(buttonClickEvent -> {
-            ProjectDto projectDto = new ProjectDto();
-            Binder<ProjectDto> binder = new Binder<>();
-            VerticalLayout content = new VerticalLayout();
-            TextField title = new TextField("Title");
-            binder.forField(title).asRequired().bind(ProjectDto::getTitle, ProjectDto::setTitle);
-            TextField description = new TextField("Description");
-            binder.forField(description).bind(ProjectDto::getDescription, ProjectDto::setDescription);
-            FormLayout gridLayout = new FormLayout();
-            gridLayout.add(title, description);
-            Button cancel = new Button("Отмена");
-            Button save = new Button("Сохранить");
-            save.setIcon(VaadinIcon.PENCIL.create());
-            content.add(gridLayout);
-            HorizontalLayout hor = new HorizontalLayout();
-            hor.add(cancel, save);
-            content.add(hor);
-            content.setAlignSelf(FlexComponent.Alignment.END, hor);
-            Dialog window = new Dialog();
-            window.add(content);
-            window.setModal(true);
-            cancel.addClickListener(new ComponentEventListener<ClickEvent<Button>>() {
-                @Override
-                public void onComponentEvent(ClickEvent<Button> buttonClickEvent) {
-                    window.close();
-                }
-            });
-            save.addClickListener(new ComponentEventListener<ClickEvent<Button>>() {
-                @Override
-                public void onComponentEvent(ClickEvent<Button> buttonClickEvent) {
-                    binder.validate();
-                    if (binder.writeBeanIfValid(projectDto)) {
-                        projectController.addProject(projectDto);
-                        window.close();
-                        projectGrid.refreshGrid();
-                    }
-                }
-            });
-            window.open();
-        });
     }
 
     private void grid(){
         projectGrid = new GridCrud<>(Project.class);
+        projectGrid.getGrid().getStyle().set("border","none");
         projectGrid.setSizeFull();
         projectGrid.getGrid().setColumns("title", "description");
         projectGrid.getGrid().addComponentColumn(project -> {
@@ -111,7 +66,7 @@ public class ProjectsView extends VerticalLayout {
         projectGrid.getGrid().addItemClickListener(new ComponentEventListener<ItemClickEvent<Project>>() {
             @Override
             public void onComponentEvent(ItemClickEvent<Project> projectItemClickEvent) {
-                UI.getCurrent().navigate(ProjectView.class,
+               UI.getCurrent().navigate(ProjectView.class,
                         projectItemClickEvent.getItem().getId().toString());
             }
         });
@@ -121,11 +76,18 @@ public class ProjectsView extends VerticalLayout {
     private Button createRemoveButton(Project project) {
         @SuppressWarnings("unchecked")
         Button button = new Button("Удалить", clickEvent -> {
-            projectController.deleteProject(project.getId());
-            projectGrid.refreshGrid();
+            ConfirmDialog confirmDialog = new ConfirmDialog(
+                    "Вы уверены что хотите удалить Проект: "+project.getTitle()+" ?",
+                    "",
+                    "Удалить",
+                    ()->{
+                        projectController.deleteProject(project.getId());
+                        projectGrid.refreshGrid();
+                    });
+            confirmDialog.open();
         });
         button.setIcon(VaadinIcon.TRASH.create());
-        button.getElement().setAttribute("theme", "error tertiary");
+        button.setId("del");
         return button;
     }
 
@@ -176,6 +138,55 @@ public class ProjectsView extends VerticalLayout {
         button.setIcon(VaadinIcon.EDIT.create());
         return button;
     }
+
+    private Button createProjectButton(){
+        Button createProject = new Button();
+        createProject.setIcon(VaadinIcon.PLUS_CIRCLE.create());
+        createProject.setText("Создать проект");
+        createProject.addClickListener(buttonClickEvent -> {
+            ProjectDto projectDto = new ProjectDto();
+            Binder<ProjectDto> binder = new Binder<>();
+            VerticalLayout content = new VerticalLayout();
+            TextField title = new TextField("Title");
+            binder.forField(title).asRequired().bind(ProjectDto::getTitle, ProjectDto::setTitle);
+            TextField description = new TextField("Description");
+            binder.forField(description).bind(ProjectDto::getDescription, ProjectDto::setDescription);
+            FormLayout gridLayout = new FormLayout();
+            gridLayout.add(title, description);
+            Button cancel = new Button("Отмена");
+            Button save = new Button("Сохранить");
+            save.setIcon(VaadinIcon.PENCIL.create());
+            content.add(gridLayout);
+            HorizontalLayout hor = new HorizontalLayout();
+            hor.add(cancel, save);
+            content.add(hor);
+            content.setAlignSelf(FlexComponent.Alignment.END, hor);
+            Dialog window = new Dialog();
+            window.add(content);
+            window.setModal(true);
+            cancel.addClickListener(new ComponentEventListener<ClickEvent<Button>>() {
+                @Override
+                public void onComponentEvent(ClickEvent<Button> buttonClickEvent) {
+                    window.close();
+                }
+            });
+            save.addClickListener(new ComponentEventListener<ClickEvent<Button>>() {
+                @Override
+                public void onComponentEvent(ClickEvent<Button> buttonClickEvent) {
+                    binder.validate();
+                    if (binder.writeBeanIfValid(projectDto)) {
+                        projectController.addProject(projectDto);
+                        window.close();
+                        projectGrid.refreshGrid();
+                    }
+                }
+            });
+            window.open();
+        });
+        return createProject;
+    }
+
+
 
 
 }
